+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-UNRELEASED
+0.20.2 (22-11-11)
 =================
 - Fixes an issue with async fixtures that are defined as methods on a test class not being rebound to the actual test instance. `#197 <https://github.com/pytest-dev/pytest-asyncio/issues/197>`_
 - Replaced usage of deprecated ``@pytest.mark.tryfirst`` with ``@pytest.hookimpl(tryfirst=True)`` `#438 <https://github.com/pytest-dev/pytest-asyncio/pull/438>`_